Added setting of reminder time

+++ app/src/main/java/com/oddhov/facebookcalendarsync/data/models/CustomTime.java
@@ -18,8 +18,8 @@ public enum CustomTime {
         this.mDisplayString = displayString;
     }
 
-    public long getTimeInMinutes() {
-        return (long) (mValue * 3600); //TODO make sure this works for 0.5 hour
+    public int getTimeInMinutes() {
+        return (int) (mValue * 3600); //TODO make sure this works for 0.5 hour
     }
 
     public String getTimeInMinutesString() {
